Renamed as_configuration_manager to as_ap_servconf
